feat:新增章节

+++ src/main/java/com/ouyanglol/crawler/model/ComicBasic.java
@@ -30,7 +30,7 @@ public class ComicBasic {
     /**
      * 连载状态 1未完结 -1完结
      */
-    private Byte serializeStatus;
+    private Integer serializeStatus;
 
     private Date createDate;
